# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: AI Exam Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /users:
    get:
      operationId: listUsers
      tags: [users]
      summary: List all users (admin)
      parameters:
        - name: role
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of users
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/User"
  /users/{id}:
    get:
      operationId: getUser
      tags: [users]
      summary: Get user by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: User
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    patch:
      operationId: updateUser
      tags: [users]
      summary: Update user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserUpdate"
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/User"
    delete:
      operationId: deleteUser
      tags: [users]
      summary: Delete user
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted

  # EXAMS
...
  /violations:
    post:
      operationId: reportViolation
      tags: [violations]
      summary: Report anti-cheat violation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ViolationInput"
      responses:
        "201":
          description: Violation recorded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Violation"
  /violations/{attemptId}:
    get:
      operationId: getViolations
      tags: [violations]
      summary: Get violations for an attempt
      parameters:
        - name: attemptId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Violations
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Violation"

  # FEEDS
...
components:
  schemas:
...
    User:
      type: object
      required: [id, name, email, role, createdAt]
      properties:
        id:
          type: integer
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, STUDENT]
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: ["string", "null"]
          format: date-time
    UserUpdate:
      type: object
      properties:
        name:
          type: string
        email:
          type: string
        role:
          type: string
          enum: [ADMIN, STUDENT]

    # EXAMS
...
    Violation:
      type: object
      required: [id, attemptId, violationType, count, createdAt]
      properties:
        id:
          type: integer
        attemptId:
          type: integer
        violationType:
          type: string
          enum: [TAB_SWITCH, FULLSCREEN_EXIT, WINDOW_BLUR, COPY_PASTE, RIGHT_CLICK, DEV_TOOLS]
        count:
          type: integer
        autoSubmitted:
          type: boolean
        createdAt:
          type: string
          format: date-time
    ViolationInput:
      type: object
      required: [attemptId, violationType]
      properties:
        attemptId:
          type: integer
        violationType:
          type: string
          enum: [TAB_SWITCH, FULLSCREEN_EXIT, WINDOW_BLUR, COPY_PASTE, RIGHT_CLICK, DEV_TOOLS]
        autoSubmit:
          type: boolean

    # DASHBOARD
...