# lib/api-zod/src/generated/types/violationInput.ts
/**
 * Generated by orval v8.12.3 🍺
 * Do not edit manually.
 * Api
 * AI Exam Platform API
 * OpenAPI spec version: 0.1.0
 */
import type { ViolationInputViolationType } from './violationInputViolationType';

export interface ViolationInput {
  attemptId: number;
  violationType: ViolationInputViolationType;
  autoSubmit?: boolean;
}